requirements.txt
affine==2.4.0
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
Cartopy==0.24.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
contextily==1.6.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
geographiclib==2.0
geopandas==1.1.0
geoplot==0.5.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
mapclassify==2.9.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mercantile==1.2.1
narwhals==1.43.0
networkx==3.5
numpy==2.3.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
referencing==0.36.2
requests==2.32.4
rioxarray==0.19.0
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.1
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
xarray==2025.6.1
xyzservices==2025.4.0